feat: Connect from an oleobject using COM connection (VB 6, Powerbuilder)
To connect from an oleobject from VB 6 or Powerbuilder, we need a
constructor without parameters since it is not possible to parameterize
it.

In addition, COM connections are not compatible with generic collections,
throws an exception "function name not found".

This can be solved by eliminating the generic collections, since it would
not affect the current functionality, and creating a constructor without
parameters. Subsequently, the QRCodeData object that is currently assigned
in the constructor with parameters, would be added manually with
SetQRCodeData function

Resolves: #108