Adding "named constructors" for QRCoder.PayloadGenerator.SwissQrCode.Contact
When trying to use the Contact constructor for a "combined address" from C++/CLR, there was an ambiguous call error (because it takes four strings and the constructor of a "structured address" takes also four string and two more strings that default to null).

Marked the existing two public constructors as obsolete.